--- a/packages/fns/tests/adjust.test.js
+++ b/packages/fns/tests/adjust.test.js
@@ -1,3 +1,5 @@
+import { describe, expect, test } from '@jest/globals';
+
 import adjust from '../src/adjust';
 
 const add = v => v + 1;