--- a/packages/fns/tests/adjust.test.js
+++ b/packages/fns/tests/adjust.test.js
@@ -1,4 +1,4 @@
-import { describe, expect, test } from '@jest/globals';
+import { describe, expect, test } from 'vitest';
 
 import adjust from '../src/adjust';