--- a/packages/fns/tests/capitalize.test.js
+++ b/packages/fns/tests/capitalize.test.js
@@ -1,4 +1,4 @@
-import { describe, expect, test } from '@jest/globals';
+import { describe, expect, test } from 'vitest';
 
 import capitalize from '../src/capitalize';