--- a/packages/fns/tests/castArray.test.js
+++ b/packages/fns/tests/castArray.test.js
@@ -1,3 +1,5 @@
+import { describe, expect, test } from '@jest/globals';
+
 import castArray from '../src/castArray';
 
 describe('castArray', () => {