--- a/packages/fns/tests/castArray.test.js
+++ b/packages/fns/tests/castArray.test.js
@@ -1,4 +1,4 @@
-import { describe, expect, test } from '@jest/globals';
+import { describe, expect, test } from 'vitest';
 
 import castArray from '../src/castArray';