--- a/packages/fns/tests/compose.test.js
+++ b/packages/fns/tests/compose.test.js
@@ -1,3 +1,5 @@
+import { describe, expect, test } from '@jest/globals';
+
 import compose from '../src/compose';
 
 describe('compose', () => {