--- a/packages/fns/tests/compose.test.js
+++ b/packages/fns/tests/compose.test.js
@@ -1,4 +1,4 @@
-import { describe, expect, test } from '@jest/globals';
+import { describe, expect, test } from 'vitest';
 
 import compose from '../src/compose';