--- a/packages/fns/tests/dropLast.test.js
+++ b/packages/fns/tests/dropLast.test.js
@@ -1,3 +1,5 @@
+import { describe, expect, test } from '@jest/globals';
+
 import dropLast from '../src/dropLast';
 
 describe('dropLast', () => {