--- a/packages/fns/tests/dropLast.test.js
+++ b/packages/fns/tests/dropLast.test.js
@@ -1,4 +1,4 @@
-import { describe, expect, test } from '@jest/globals';
+import { describe, expect, test } from 'vitest';
 
 import dropLast from '../src/dropLast';