--- a/packages/fns/tests/evolve.test.js
+++ b/packages/fns/tests/evolve.test.js
@@ -1,4 +1,4 @@
-import { describe, expect, test } from '@jest/globals';
+import { describe, expect, test } from 'vitest';
 
 import evolve from '../src/evolve';