--- a/packages/fns/tests/get.test.js
+++ b/packages/fns/tests/get.test.js
@@ -1,3 +1,5 @@
+import { describe, expect, test } from '@jest/globals';
+
 import get from '../src/get';
 
 describe('get', () => {