--- a/packages/fns/tests/get.test.js
+++ b/packages/fns/tests/get.test.js
@@ -1,4 +1,4 @@
-import { describe, expect, test } from '@jest/globals';
+import { describe, expect, test } from 'vitest';
 
 import get from '../src/get';