--- a/packages/fns/tests/isNil.test.js
+++ b/packages/fns/tests/isNil.test.js
@@ -1,4 +1,5 @@
 /* eslint-disable no-void */
+import { describe, expect, test } from '@jest/globals';
 
 import isNil from '../src/isNil';