--- a/packages/fns/tests/last.test.js
+++ b/packages/fns/tests/last.test.js
@@ -1,3 +1,5 @@
+import { describe, expect, test } from '@jest/globals';
+
 import last from '../src/last';
 
 describe('last', () => {