--- a/packages/fns/tests/last.test.js
+++ b/packages/fns/tests/last.test.js
@@ -1,4 +1,4 @@
-import { describe, expect, test } from '@jest/globals';
+import { describe, expect, test } from 'vitest';
 
 import last from '../src/last';