--- a/packages/fns/tests/matchPercent.test.js
+++ b/packages/fns/tests/matchPercent.test.js
@@ -1,3 +1,5 @@
+import { describe, expect, test } from '@jest/globals';
+
 import matchPercent from '../src/matchPercent';
 
 describe('match percent', () => {