--- a/packages/fns/tests/matchPercent.test.js
+++ b/packages/fns/tests/matchPercent.test.js
@@ -1,4 +1,4 @@
-import { describe, expect, test } from '@jest/globals';
+import { describe, expect, test } from 'vitest';
 
 import matchPercent from '../src/matchPercent';