--- a/packages/fns/tests/omit.test.js
+++ b/packages/fns/tests/omit.test.js
@@ -1,3 +1,5 @@
+import { describe, expect, test } from '@jest/globals';
+
 import omit from '../src/omit';
 
 describe('omit', () => {