--- a/packages/fns/tests/omit.test.js
+++ b/packages/fns/tests/omit.test.js
@@ -1,4 +1,4 @@
-import { describe, expect, test } from '@jest/globals';
+import { describe, expect, test } from 'vitest';
 
 import omit from '../src/omit';