--- a/packages/fns/tests/pick.test.js
+++ b/packages/fns/tests/pick.test.js
@@ -1,3 +1,5 @@
+import { describe, expect, test } from '@jest/globals';
+
 import pick from '../src/pick';
 
 describe('pick', () => {