--- a/packages/fns/tests/pick.test.js
+++ b/packages/fns/tests/pick.test.js
@@ -1,4 +1,4 @@
-import { describe, expect, test } from '@jest/globals';
+import { describe, expect, test } from 'vitest';
 
 import pick from '../src/pick';