--- a/packages/fns/tests/reverse.test.js
+++ b/packages/fns/tests/reverse.test.js
@@ -1,3 +1,5 @@
+import { describe, expect, test } from '@jest/globals';
+
 import reverse from '../src/reverse';
 
 describe('reverse', () => {