--- a/packages/fns/tests/reverse.test.js
+++ b/packages/fns/tests/reverse.test.js
@@ -1,4 +1,4 @@
-import { describe, expect, test } from '@jest/globals';
+import { describe, expect, test } from 'vitest';
 
 import reverse from '../src/reverse';