--- a/packages/fns/tests/upperFirst.test.js
+++ b/packages/fns/tests/upperFirst.test.js
@@ -1,4 +1,4 @@
-import { describe, expect, test } from '@jest/globals';
+import { describe, expect, test } from 'vitest';
 
 import upperFirst from '../src/upperFirst';