--- a/packages/font/setupTests.js
+++ b/packages/font/setupTests.js
@@ -1,15 +1,12 @@
 /* eslint-disable */
-import { jest } from '@jest/globals';
-import { createRequire } from 'module';
+import { vi } from 'vitest';
+import createFetchMock from 'vitest-fetch-mock';
 
 import '../../polyfills';
 
-const require = createRequire(import.meta.url);
+const fetchMocker = createFetchMock(vi);
+fetchMocker.enableMocks();
 
 global.BROWSER = false;
 
-const customGlobal = global;
-customGlobal.fetch = require('jest-fetch-mock');
-customGlobal.fetchMock = customGlobal.fetch;
-
-jest.setMock('cross-fetch', fetch);
+vi.mock('cross-fetch', () => ({ default: global.fetch }));