--- a/packages/image/jest.config.js
+++ b/packages/image/jest.config.js
@@ -1,5 +1,0 @@
-export default {
-  automock: false,
-  testRegex: 'tests/.*?(test)\\.js$',
-  setupFiles: ['<rootDir>setupTests.js'],
-};