--- a/packages/image/tests/cache.test.js
+++ b/packages/image/tests/cache.test.js
@@ -1,4 +1,4 @@
-import { describe, expect, test } from '@jest/globals';
+import { describe, expect, test } from 'vitest';
 
 import createCache from '../src/cache';