--- a/packages/image/tests/resolve.test.js
+++ b/packages/image/tests/resolve.test.js
@@ -1,3 +1,4 @@
+import { beforeEach, describe, expect, test } from '@jest/globals';
 import fs from 'fs';
 import path from 'path';
 import url from 'url';