--- a/packages/image/vitest.config.js
+++ b/packages/image/vitest.config.js
@@ -1,0 +1,11 @@
+// false positive on import/no-unresolved
+// eslint-disable-next-line import/no-unresolved
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    setupFiles: ['vitest.setup.js'],
+    include: ['tests/*.{test,spec}.?(c|m)[jt]s?(x)'],
+    watch: false,
+  },
+});