--- a/packages/layout/setupTests.js
+++ b/packages/layout/setupTests.js
@@ -1,14 +1,12 @@
 /* eslint-disable */
-import { createRequire } from 'module';
+import { vi } from 'vitest';
+import createFetchMock from 'vitest-fetch-mock';
 
 import '../../polyfills';
 
-const require = createRequire(import.meta.url);
+const fetchMocker = createFetchMock(vi);
+fetchMocker.enableMocks();
 
 global.BROWSER = false;
 
-const customGlobal = global;
-customGlobal.fetch = require('jest-fetch-mock');
-customGlobal.fetchMock = customGlobal.fetch;
-
-jest.setMock('cross-fetch', fetch);
+vi.mock('cross-fetch', () => ({ default: global.fetch }));