--- a/packages/layout/tests/image/getSource.test.js
+++ b/packages/layout/tests/image/getSource.test.js
@@ -1,4 +1,4 @@
-import { describe, expect, test } from '@jest/globals';
+import { describe, expect, test } from 'vitest';
 
 import getSource from '../../src/image/getSource';