--- a/packages/layout/tests/image/resolveSource.test.js
+++ b/packages/layout/tests/image/resolveSource.test.js
@@ -1,4 +1,4 @@
-import { describe, expect, it } from '@jest/globals';
+import { describe, expect, it } from 'vitest';
 
 import resolveSource from '../../src/image/resolveSource';