--- a/packages/layout/tests/node/getBorderWidth.test.js
+++ b/packages/layout/tests/node/getBorderWidth.test.js
@@ -1,3 +1,5 @@
+import { describe, expect, test } from '@jest/globals';
+
 import * as Yoga from 'yoga-layout';
 
 import getBorderWidth from '../../src/node/getBorderWidth';