--- a/packages/layout/tests/node/getBorderWidth.test.js
+++ b/packages/layout/tests/node/getBorderWidth.test.js
@@ -1,4 +1,4 @@
-import { describe, expect, test } from '@jest/globals';
+import { describe, expect, test } from 'vitest';
 
 import * as Yoga from 'yoga-layout';