--- a/packages/layout/tests/node/getDimension.test.js
+++ b/packages/layout/tests/node/getDimension.test.js
@@ -1,4 +1,4 @@
-import { describe, expect, test } from '@jest/globals';
+import { describe, expect, test } from 'vitest';
 
 import getDimension from '../../src/node/getDimension';