--- a/packages/layout/tests/node/getMargin.test.js
+++ b/packages/layout/tests/node/getMargin.test.js
@@ -1,4 +1,4 @@
-import { describe, expect, test } from '@jest/globals';
+import { describe, expect, test } from 'vitest';
 
 import * as Yoga from 'yoga-layout';