--- a/packages/layout/tests/node/getOrigin.test.js
+++ b/packages/layout/tests/node/getOrigin.test.js
@@ -1,4 +1,4 @@
-import { describe, expect, test } from '@jest/globals';
+import { describe, expect, test } from 'vitest';
 
 import getOrigin from '../../src/node/getOrigin';