--- a/packages/layout/tests/node/getPadding.test.js
+++ b/packages/layout/tests/node/getPadding.test.js
@@ -1,3 +1,5 @@
+import { describe, expect, test } from '@jest/globals';
+
 import * as Yoga from 'yoga-layout';
 
 import getPadding from '../../src/node/getPadding';