--- a/packages/layout/tests/node/getPosition.test.js
+++ b/packages/layout/tests/node/getPosition.test.js
@@ -1,4 +1,4 @@
-import { describe, expect, test } from '@jest/globals';
+import { describe, expect, test } from 'vitest';
 
 import getPosition from '../../src/node/getPosition';