--- a/packages/layout/tests/node/removePaddings.test.js
+++ b/packages/layout/tests/node/removePaddings.test.js
@@ -1,3 +1,5 @@
+import { describe, expect, test } from '@jest/globals';
+
 import removePaddings from '../../src/node/removePaddings';
 
 describe('node removePaddings', () => {