--- a/packages/layout/tests/node/removePaddings.test.js
+++ b/packages/layout/tests/node/removePaddings.test.js
@@ -1,4 +1,4 @@
-import { describe, expect, test } from '@jest/globals';
+import { describe, expect, test } from 'vitest';
 
 import removePaddings from '../../src/node/removePaddings';