--- a/packages/layout/tests/node/setAlignContent.test.js
+++ b/packages/layout/tests/node/setAlignContent.test.js
@@ -1,11 +1,11 @@
-import { beforeEach, describe, expect, jest, test } from '@jest/globals';
+import { beforeEach, describe, expect, test, vi } from 'vitest';
 
 import * as Yoga from 'yoga-layout';
 
 import setAlignContent from '../../src/node/setAlignContent';
 
 describe('node setAlignContent', () => {
-  const mock = jest.fn();
+  const mock = vi.fn();
   const node = { yogaNode: { setAlignContent: mock } };
 
   beforeEach(() => {