--- a/packages/layout/tests/node/setAlignItems.test.js
+++ b/packages/layout/tests/node/setAlignItems.test.js
@@ -1,4 +1,5 @@
-import { jest } from '@jest/globals';
+import { beforeEach, describe, expect, jest, test } from '@jest/globals';
+
 import * as Yoga from 'yoga-layout';
 
 import setAlignItems from '../../src/node/setAlignItems';