--- a/packages/layout/tests/node/setAlignItems.test.js
+++ b/packages/layout/tests/node/setAlignItems.test.js
@@ -1,11 +1,11 @@
-import { beforeEach, describe, expect, jest, test } from '@jest/globals';
+import { beforeEach, describe, expect, test, vi } from 'vitest';
 
 import * as Yoga from 'yoga-layout';
 
 import setAlignItems from '../../src/node/setAlignItems';
 
 describe('node setAlignItems', () => {
-  const mock = jest.fn();
+  const mock = vi.fn();
   const node = { yogaNode: { setAlignItems: mock } };
 
   beforeEach(() => {