--- a/packages/layout/tests/node/setAlignSelf.test.js
+++ b/packages/layout/tests/node/setAlignSelf.test.js
@@ -1,4 +1,5 @@
-import { jest } from '@jest/globals';
+import { beforeEach, describe, expect, jest, test } from '@jest/globals';
+
 import * as Yoga from 'yoga-layout';
 
 import setAlignSelf from '../../src/node/setAlignSelf';