--- a/packages/layout/tests/node/setAspectRatio.test.js
+++ b/packages/layout/tests/node/setAspectRatio.test.js
@@ -1,4 +1,4 @@
-import { jest } from '@jest/globals';
+import { beforeEach, describe, expect, jest, test } from '@jest/globals';
 import setAspectRatio from '../../src/node/setAspectRatio';
 
 describe('node setAspectRatio', () => {