--- a/packages/layout/tests/node/setFlexBasis.test.js
+++ b/packages/layout/tests/node/setFlexBasis.test.js
@@ -1,8 +1,8 @@
-import { beforeEach, describe, expect, jest, test } from '@jest/globals';
+import { beforeEach, describe, expect, test, vi } from 'vitest';
 import setFlexBasis from '../../src/node/setFlexBasis';
 
 describe('node setFlexBasis', () => {
-  const mock = jest.fn();
+  const mock = vi.fn();
   const node = { yogaNode: { setFlexBasis: mock } };
 
   beforeEach(() => {