--- a/packages/layout/tests/node/setFlexDirection.test.js
+++ b/packages/layout/tests/node/setFlexDirection.test.js
@@ -1,4 +1,5 @@
-import { jest } from '@jest/globals';
+import { beforeEach, describe, expect, jest, test } from '@jest/globals';
+
 import * as Yoga from 'yoga-layout';
 
 import setFlexDirection from '../../src/node/setFlexDirection';