--- a/packages/layout/tests/node/setFlexDirection.test.js
+++ b/packages/layout/tests/node/setFlexDirection.test.js
@@ -1,11 +1,11 @@
-import { beforeEach, describe, expect, jest, test } from '@jest/globals';
+import { beforeEach, describe, expect, test, vi } from 'vitest';
 
 import * as Yoga from 'yoga-layout';
 
 import setFlexDirection from '../../src/node/setFlexDirection';
 
 describe('node setFlexDirection', () => {
-  const mock = jest.fn();
+  const mock = vi.fn();
   const node = { yogaNode: { setFlexDirection: mock } };
 
   beforeEach(() => {