--- a/packages/layout/tests/node/setFlexShrink.test.js
+++ b/packages/layout/tests/node/setFlexShrink.test.js
@@ -1,4 +1,4 @@
-import { jest } from '@jest/globals';
+import { beforeEach, describe, expect, jest, test } from '@jest/globals';
 import setFlexShrink from '../../src/node/setFlexShrink';
 
 describe('node setFlexShrink', () => {