--- a/packages/layout/tests/node/setFlexShrink.test.js
+++ b/packages/layout/tests/node/setFlexShrink.test.js
@@ -1,8 +1,8 @@
-import { beforeEach, describe, expect, jest, test } from '@jest/globals';
+import { beforeEach, describe, expect, test, vi } from 'vitest';
 import setFlexShrink from '../../src/node/setFlexShrink';
 
 describe('node setFlexShrink', () => {
-  const mock = jest.fn();
+  const mock = vi.fn();
   const node = { yogaNode: { setFlexShrink: mock } };
 
   beforeEach(() => {