--- a/packages/layout/tests/node/setJustifyContent.test.js
+++ b/packages/layout/tests/node/setJustifyContent.test.js
@@ -1,11 +1,11 @@
-import { beforeEach, describe, expect, jest, test } from '@jest/globals';
+import { beforeEach, describe, expect, test, vi } from 'vitest';
 
 import * as Yoga from 'yoga-layout';
 
 import setJustifyContent from '../../src/node/setJustifyContent';
 
 describe('node setJustifyContent', () => {
-  const mock = jest.fn();
+  const mock = vi.fn();
   const node = { yogaNode: { setJustifyContent: mock } };
 
   beforeEach(() => {