--- a/packages/layout/tests/node/setPadding.test.js
+++ b/packages/layout/tests/node/setPadding.test.js
@@ -1,4 +1,4 @@
-import { beforeEach, describe, expect, jest, test } from '@jest/globals';
+import { beforeEach, describe, expect, test, vi } from 'vitest';
 
 import * as Yoga from 'yoga-layout';
 
@@ -10,8 +10,8 @@ import setPadding, {
 } from '../../src/node/setPadding';
 
 describe('node setPadding', () => {
-  const mock = jest.fn();
-  const mockPercent = jest.fn();
+  const mock = vi.fn();
+  const mockPercent = vi.fn();
 
   const node = {
     yogaNode: { setPadding: mock, setPaddingPercent: mockPercent },