--- a/packages/renderer/jest.config.js
+++ b/packages/renderer/jest.config.js
@@ -1,8 +1,0 @@
-export default {
-  testRegex: 'tests/.*?(test)\\.js$',
-  setupFilesAfterEnv: ['<rootDir>/setup.jest.js'],
-  transformIgnorePatterns: [
-    '/node_modules/',
-    '<rootDir>/tests/environment/jsdom.js',
-  ],
-};