--- a/packages/renderer/setup.jest.js
+++ b/packages/renderer/setup.jest.js
@@ -1,3 +1,4 @@
+import { expect } from '@jest/globals';
 import path from 'path';
 import url from 'url';
 import jestImageSnapshot from 'jest-image-snapshot';