--- a/packages/textkit/src/utils/stringFromCodePoints.js
+++ b/packages/textkit/src/utils/stringFromCodePoints.js
@@ -4,6 +4,7 @@
  * @param {number[]} codePoints points
  * @returns {string} string
  */
-const stringFromCodePoints = codePoints => String.fromCodePoint(...codePoints);
+const stringFromCodePoints = codePoints =>
+  String.fromCodePoint(...(codePoints || []));
 
 export default stringFromCodePoints;