--- a/packages/types/context.d.ts
+++ b/packages/types/context.d.ts
@@ -1,20 +1,19 @@
-
-// instance.registerFont = jest.fn().mockReturnValue(instance);
-// instance.note = jest.fn().mockReturnValue(instance);
-// instance.rotate = jest.fn().mockReturnValue(instance);
-// instance.scale = jest.fn().mockReturnValue(instance);
-// instance.translate = jest.fn().mockReturnValue(instance);
-// instance.link = jest.fn().mockReturnValue(instance);
-// instance.goTo = jest.fn().mockReturnValue(instance);
-// instance.addNamedDestination = jest.fn().mockReturnValue(instance);
-// instance.clip = jest.fn().mockReturnValue(instance);
-// instance.closePath = jest.fn().mockReturnValue(instance);
-// instance.path = jest.fn().mockReturnValue(instance);
-// instance.radialGradient = jest.fn().mockReturnValue(instance);
-// instance.linearGradient = jest.fn().mockReturnValue(instance);
-// instance.fontSize = jest.fn().mockReturnValue(instance);
-// instance.text = jest.fn().mockReturnValue(instance);
-// instance.font = jest.fn().mockReturnValue(instance);
+// instance.registerFont = vi.fn().mockReturnValue(instance);
+// instance.note = vi.fn().mockReturnValue(instance);
+// instance.rotate = vi.fn().mockReturnValue(instance);
+// instance.scale = vi.fn().mockReturnValue(instance);
+// instance.translate = vi.fn().mockReturnValue(instance);
+// instance.link = vi.fn().mockReturnValue(instance);
+// instance.goTo = vi.fn().mockReturnValue(instance);
+// instance.addNamedDestination = vi.fn().mockReturnValue(instance);
+// instance.clip = vi.fn().mockReturnValue(instance);
+// instance.closePath = vi.fn().mockReturnValue(instance);
+// instance.path = vi.fn().mockReturnValue(instance);
+// instance.radialGradient = vi.fn().mockReturnValue(instance);
+// instance.linearGradient = vi.fn().mockReturnValue(instance);
+// instance.fontSize = vi.fn().mockReturnValue(instance);
+// instance.text = vi.fn().mockReturnValue(instance);
+// instance.font = vi.fn().mockReturnValue(instance);
 
 interface DocumentInfo {
   Producer?: string;
@@ -66,5 +65,12 @@ export interface Context {
   quadraticCurveTo(cpx: number, cpy: number, x: number, y: number): this;
   image(src: any, x?: number, y?: number, options?: ImageOption): Context;
   roundedRect(x: number, y: number, w: number, h: number, r?: number): this;
-  bezierCurveTo(cp1x: number, cp1y: number, cp2x: number, cp2y: number, x: number, y: number): this;
+  bezierCurveTo(
+    cp1x: number,
+    cp1y: number,
+    cp2x: number,
+    cp2y: number,
+    x: number,
+    y: number,
+  ): this;
 }