--- a/vitest.workspace.js
+++ b/vitest.workspace.js
@@ -1,0 +1,13 @@
+export default [
+  'packages/fns',
+  'packages/font',
+  'packages/image',
+  'packages/render',
+  'packages/layout',
+  'packages/svgkit',
+  'packages/textkit',
+  'packages/renderer/vitest.config.js',
+  'packages/renderer/vitest.browser.config.js',
+  'packages/stylesheet',
+  'packages/primitives',
+];