Requires #2409.

~~Requires #2494.~~

Migrates the entire monorepo from Jest to Vitest.