@diegomura 2024-01-22
@wojtekmaj Just to understand this better (not familiarized with vitest), the benefit of this is better support for ESM and more performant? About the latter I'm not so concern as the entire testsuite runs in like 6sec. Also curious how tight this lib is ti vite, considering we don't use it here