@diegomura 2024-01-22
Actually, about performance, I tested this branch and tests take 50% more to run (9s). It's not bad either s not concerned, but still curious :)