@wojtekmaj 2024-01-22
Hmmm, interesting. You're right that the overall run time has increased a bit, and this is while the individual tests run faster:

Before:
 PASS  packages/renderer/tests/resume.test.js (6.884 s)

After:
 ✓ |renderer| tests/resume.test.jsx  (3 tests) 3767ms

Before:
 PASS  packages/renderer/tests/pageWrap.test.js (7.924 s)

After:
 ✓ |renderer| tests/pageWrap.test.jsx  (1 test) 3553ms

This makes me convinced that while there is probably something I could still tweak, the more tests and the more complex they are, the faster Vitest will be in relation to Jest.