@wojtekmaj 2024-01-22
Now it's 35% faster than Jest on CI! ;)